frontend/telegram-bot/requirements.txt
python-telegram-bot>=21.0,<23.0
httpx>=0.27,<1.0

